+++ DogeCoiner.Data/Dtos/Transaction.cs
@@ -4,8 +4,8 @@ namespace DogeCoiner.Data.Dtos
 {
     public class Transaction
     {
-        public int TransactionId { get; set; }
-        public int PortfolioId { get; set; }
+        public long TransactionId { get; set; }
+        public long PortfolioId { get; set; }
         public string Symbol { get; set; }
         public string TransactionType { get; set; }
         public decimal Quantity { get; set; }
